+++ otterdog/models/environment.py
@@ -22,6 +22,7 @@ from otterdog.models import (
     LivePatchType,
     LivePatchContext,
     LivePatchHandler,
+    PatchContext,
 )
 from otterdog.providers.github import GitHubProvider
 from otterdog.utils import (
@@ -188,6 +189,7 @@ class Environment(ModelObject):
         self,
         printer: IndentingPrinter,
         jsonnet_config: JsonnetConfig,
+        context: PatchContext,
         extend: bool,
         default_object: ModelObject,
     ) -> None: